+++ src/MonoidK.ts
@@ -1,10 +1,7 @@
 /**
- * The `Alternative` type class extends the `SemigroupK` type class with a value that should be the left and right identity for `alt`.
+ * TODO: description
  *
- * It is similar to `Monoid`, except that it applies to types of kind `* -> *`, like `Array` or `Option`, rather than
- * concrete types like `string` or `number`.
- *
- * `Alternative` instances should satisfy the following laws in addition to the `SemigroupK` laws:
+ * `MonoidK` instances should satisfy the following laws in addition to the `SemigroupK` laws:
  *
  * 1. Left identity: `zero |> alt(() => fa) <-> fa`
  * 2. Right identity: `fa |> alt(() => zero) <-> fa`
@@ -27,7 +24,7 @@ import type { Pointed } from './Pointed'
  * @category type classes
  * @since 3.0.0
  */
-export interface Alternative<F extends HKT> extends SemigroupK<F> {
+export interface MonoidK<F extends HKT> extends SemigroupK<F> {
   readonly zero: <S, R = unknown, W = never, E = never, A = never>() => Kind<F, S, R, W, E, A>
 }
 
@@ -40,7 +37,7 @@ export interface Alternative<F extends HKT> extends SemigroupK<F> {
  * @since 3.0.0
  */
 export const guard =
-  <F extends HKT>(F: Alternative<F>, P: Pointed<F>) =>
+  <F extends HKT>(F: MonoidK<F>, P: Pointed<F>) =>
   <S, R = unknown, W = never, E = never>(b: boolean): Kind<F, S, R, W, E, void> =>
     b ? P.of(undefined) : F.zero()
 
@@ -52,6 +49,6 @@ export const guard =
  * @since 3.0.0
  */
 export const altAll = <F extends HKT>(
-  F: Alternative<F>
+  F: MonoidK<F>
 ): (<S, R, W, E, A>(as: ReadonlyArray<Kind<F, S, R, W, E, A>>) => Kind<F, S, R, W, E, A>) =>
   semigroupK.altAll(F)(F.zero())