+++ src/ReadonlyArray.ts
@@ -2,7 +2,7 @@
  * @since 3.0.0
  */
 import type * as semigroupK from './SemigroupK'
-import * as alternative from './Alternative'
+import * as monoidK from './MonoidK'
 import type * as applicative from './Applicative'
 import * as apply from './Apply'
 import * as chainable from './Chainable'
@@ -1245,7 +1245,7 @@ export const difference = <A>(E: Eq<A>): Magma<ReadonlyArray<A>>['concat'] => {
 export const of: <A>(a: A) => ReadonlyArray<A> = readonlyNonEmptyArray.of
 
 /**
- * @category Alternative
+ * @category MonoidK
  * @since 3.0.0
  */
 export const zero: <A>() => ReadonlyArray<A> = () => empty
@@ -1845,7 +1845,7 @@ export const SemigroupK: semigroupK.SemigroupK<ReadonlyArrayF> = {
  * @category instances
  * @since 3.0.0
  */
-export const Alternative: alternative.Alternative<ReadonlyArrayF> = {
+export const MonoidK: monoidK.MonoidK<ReadonlyArrayF> = {
   alt,
   zero
 }
@@ -1854,7 +1854,7 @@ export const Alternative: alternative.Alternative<ReadonlyArrayF> = {
  * @category constructors
  * @since 3.0.0
  */
-export const guard: (b: boolean) => ReadonlyArray<void> = /*#__PURE__*/ alternative.guard(Alternative, Pointed)
+export const guard: (b: boolean) => ReadonlyArray<void> = /*#__PURE__*/ monoidK.guard(MonoidK, Pointed)
 
 /**
  * @category instances